+++ packages/react-components/react-toast/src/components/ToastContainer/ToastContainer.types.ts
@@ -1,0 +1,36 @@
+import * as React from 'react';
+import type { ComponentProps, ComponentState, Slot } from '@fluentui/react-utilities';
+import { Announce } from '../AriaLive/AriaLive.types';
+import { Toast as VanillaToast } from '../../state';
+import { ToastContextValue } from '../../contexts/toastContext';
+import { TimerProps } from '../Timer/Timer';
+
+export type ToastContainerContextValues = {
+  toast: ToastContextValue;
+};
+
+export type ToastContainerSlots = {
+  root: NonNullable<Slot<'div'>>;
+  timer: NonNullable<Slot<TimerProps>>;
+};
+
+/**
+ * ToastContainer Props
+ */
+export type ToastContainerProps = ComponentProps<Partial<ToastContainerSlots>> &
+  VanillaToast & {
+    visible: boolean;
+    announce: Announce;
+  };
+
+/**
+ * State used in rendering ToastContainer
+ */
+export type ToastContainerState = ComponentState<ToastContainerSlots> &
+  Pick<ToastContainerProps, 'remove' | 'close' | 'updateId' | 'visible'> & {
+    transitionTimeout: number;
+    timerTimeout: number;
+    running: boolean;
+    onTransitionEntering: () => void;
+    nodeRef: React.Ref<HTMLDivElement>;
+  };