+++ packages/react-components/react-toast/src/state/vanilla/toaster.ts
@@ -48,7 +48,7 @@ export class Toaster {
       priority: 0,
       pauseOnHover: false,
       pauseOnWindowBlur: false,
-      position: 'bottom-right',
+      position: 'bottom-end',
       timeout: 3000,
     };
     this.queue = createPriorityQueue<Toast>((a, b) => {
@@ -177,6 +177,7 @@ export class Toaster {
       updateId: 0,
       toasterId,
       dispatchedAt: Date.now(),
+      data: {},
     };
 
     assignDefined<Toast>(toast, toastOptions);